perf(code-runtime): cache runner script layer
