+++ platform/backend/src/types/asset-modules.d.ts
@@ -1,0 +1,4 @@
+declare module "*.py" {
+  const content: string;
+  export default content;
+}
